dro2/requirements.txt
matplotlib>=3.0.2
mock>=2.0.0
pyzmq>=18.1.0
tensorflow==1.13.1
tensorflow-probability==0.6.0
